+++ module/compiler/index.ts
@@ -1,6 +1,5 @@
 export * from './src/util';
 export * from './src/presence';
-export * from './src/module';
 export * from './src/source';
 export * from './src/compiler';
 export * from './src/transformer/manager';
